redux config + SingIn refactor
