+++ src/Redux/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+import userReducer from "./features/user/userSlice";
+
+export const store = configureStore({
+  reducer: {
+    user: userReducer,
+  },
+});
