+++ src/components/SignIn/SignIn.jsx
@@ -1,0 +1,142 @@
+// eslint-disable-next-line no-unused-vars
+import * as React from "react";
+import {
+  Avatar,
+  Box,
+  Button,
+  Checkbox,
+  FormControlLabel,
+  Grid,
+  Link,
+  Paper,
+  TextField,
+  Typography,
+} from "@mui/material";
+import FooterSI from "./FooterSI"; //sign in footer imported as a separate component
+import FertilizeImg from "../../assets/fertilize.png";
+
+export default function SignInSide() {
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    const data = new FormData(event.currentTarget);
+    console.log({
+      email: data.get("email"),
+      password: data.get("password"),
+    });
+  };
+
+  return (
+    <Grid container component="main" sx={{ height: "100vh" }}>
+      {/* left image */}
+      <Grid
+        item
+        xs={false}
+        sm={4}
+        md={7}
+        sx={{
+          // backgroundImage: `url(${FertilizeImg})`,
+          backgroundImage:
+            "url(https://images.unsplash.com/photo-1492496913980-501348b61469?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1974&q=80)",
+          backgroundRepeat: "no-repeat",
+          backgroundColor: (t) =>
+            t.palette.mode === "light"
+              ? t.palette.grey[50]
+              : t.palette.grey[900],
+          backgroundSize: "cover",
+          backgroundPosition: "center",
+        }}
+      >
+        {/* <Box
+          xs={false}
+          sx={{
+            my: 18,
+            mx: 4,
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+        >
+          <Typography component="h2" variant="h5" color={"white"}>
+            ParquerGo
+          </Typography>
+        </Box> */}
+      </Grid>
+      {/* main form */}
+      <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
+        <Box
+          sx={{
+            my: 8,
+            mx: 4,
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+        >
+          <Avatar
+            alt="compost login icon"
+            src={FertilizeImg}
+            sx={{ width: 70, height: 70 }}
+          />
+          <Typography component="h1" variant="h5">
+            PaquerGo
+          </Typography>
+          <Typography component="p" variant="body1">
+            Conectando Amigos y Paquerxs
+          </Typography>
+          <Box
+            component="form"
+            noValidate
+            onSubmit={handleSubmit}
+            sx={{ mt: 1 }}
+          >
+            <TextField
+              margin="normal"
+              required
+              fullWidth
+              id="email"
+              label="Email"
+              name="email"
+              autoComplete="email"
+              autoFocus
+            />
+            <TextField
+              margin="normal"
+              required
+              fullWidth
+              name="password"
+              label="Contraseña"
+              type="password"
+              id="password"
+              autoComplete="current-password"
+            />
+            <FormControlLabel
+              control={<Checkbox value="remember" color="primary" />}
+              label="Recordar mis datos"
+            />
+            <Button
+              type="submit"
+              fullWidth
+              variant="contained"
+              sx={{ mt: 3, mb: 2 }}
+            >
+              Ingresar
+            </Button>
+            <Grid container>
+              {/* <Grid item xs>
+                <Link href="#" variant="body2">
+                  Forgot password?
+                </Link>
+              </Grid> */}
+              <Grid item textAlign={"center"} width={"100%"}>
+                <Link href="#" variant="body2">
+                  {"No tienes cuenta? Crea tu usuario"}
+                </Link>
+              </Grid>
+            </Grid>
+            <FooterSI sx={{ mt: 3, fontSize: "10px" }} />
+          </Box>
+        </Box>
+      </Grid>
+    </Grid>
+  );
+}
