--- src/components/SignInSide/index.js
@@ -1,3 +1,0 @@
-import SignInSide from "./SignInSide";
-
-export { SignInSide };
